move stories outside of src
