+++ package.json
@@ -123,6 +123,7 @@
     "eslint-plugin-unicorn": "^61.0.0",
     "eslint-plugin-unused-imports": "^4.2.0",
     "globals": "^16.3.0",
+    "jiti": "^2.5.1",
     "jsdom": "^26.1.0",
     "postcss": "^8.4.32",
     "prettier": "^3.6.2",
